+++ src/reversion/revisions.py
@@ -4,33 +4,32 @@
 try:
     from functools import wraps
 except ImportError:
-    from django.utils.functional import wraps  # Python 2.3, 2.4 fallback.
+    from django.utils.functional import wraps  # Python 2.4 fallback.
 
 import operator
 from threading import local
 
 from django.contrib.contenttypes.models import ContentType
 from django.core import serializers
+from django.core.exceptions import ObjectDoesNotExist
 from django.db import models
 from django.db.models import Q, Max
 from django.db.models.query import QuerySet
-from django.db.models.signals import post_save
+from django.db.models.signals import post_save, pre_delete
 
 from reversion.errors import RevisionManagementError, RegistrationError
-from reversion.models import Revision, Version
-from reversion.storage import VersionFileStorageWrapper
+from reversion.models import Revision, Version, VERSION_ADD, VERSION_CHANGE, VERSION_DELETE
 
 
 class RegistrationInfo(object):
     
     """Stored registration information about a model."""
     
-    __slots__ = "fields", "file_fields", "follow", "format",
+    __slots__ = "fields", "follow", "format",
     
-    def __init__(self, fields, file_fields, follow, format):
+    def __init__(self, fields, follow, format):
         """Initializes the registration info."""
         self.fields = fields
-        self.file_fields = file_fields
         self.follow = follow
         self.format = format
 
@@ -45,7 +44,7 @@ class RevisionState(local):
     
     def clear(self):
         """Puts the revision manager back into its default state."""
-        self.objects = set()
+        self.objects = {}
         self.user = None
         self.comment = ""
         self.depth = 0
@@ -91,19 +90,13 @@ class RevisionManager(object):
         if fields is None:
             fields = [field.name for field in local_fields]
         fields = tuple(fields)
-        # Calculate serializable model file fields.
-        file_fields = []
-        for field in local_fields:
-            if isinstance(field, models.FileField) and field.name in fields:
-                field.storage = VersionFileStorageWrapper(field.storage)
-                file_fields.append(field)
-        file_fields = tuple(file_fields)
         # Register the generated registration information.
         follow = tuple(follow)
-        registration_info = RegistrationInfo(fields, file_fields, follow, format)
+        registration_info = RegistrationInfo(fields, follow, format)
         self._registry[model_class] = registration_info
         # Connect to the post save signal of the model.
         post_save.connect(self.post_save_receiver, model_class)
+        pre_delete.connect(self.pre_delete_receiver, model_class)
     
     def get_registration_info(self, model_class):
         """Returns the registration information for the given model class."""
@@ -121,9 +114,8 @@ class RevisionManager(object):
         except KeyError:
             raise RegistrationError, "%r has not been registered with Reversion." % model_class
         else:
-            for field in registration_info.file_fields:
-                field.storage = field.storage.wrapped_storage
             post_save.disconnect(self.post_save_receiver, model_class)
+            pre_delete.disconnect(self.pre_delete_receiver, model_class)
     
     # Low-level revision management methods.
     
@@ -146,10 +138,10 @@ class RevisionManager(object):
         if not self.is_active():
             raise RevisionManagementError, "There is no active revision for this thread."
     
-    def add(self, obj):
+    def add(self, obj, type_flag=VERSION_CHANGE):
         """Adds an object to the current revision."""
         self.assert_active()
-        self._state.objects.add(obj)
+        self._state.objects[obj] = self.get_version_data(obj, type_flag)
         
     def set_user(self, user):
         """Sets the user for the current revision"""
@@ -207,18 +199,18 @@ class RevisionManager(object):
         """Checks whether this revision is invalid."""
         return self._state.is_invalid
         
-    def follow_relationships(self, object_set):
+    def follow_relationships(self, object_dict):
         """
         Follows all the registered relationships in the given set of models to
         yield a set containing the original models plus all their related
         models.
         """
-        result_set = set()
+        result_dict = {}
         def _follow_relationships(obj):
             # Prevent recursion.
-            if obj in result_set:
+            if obj in result_dict or obj.pk is None:  # This last condition is because during a delete action the parent field for a subclassing model will be set to None.
                 return
-            result_set.add(obj)
+            result_dict[obj] = self.get_version_data(obj, VERSION_CHANGE)
             # Follow relations.
             registration_info = self.get_registration_info(obj.__class__)
             for relationship in registration_info.follow:
@@ -234,7 +226,7 @@ class RevisionManager(object):
                 # Get the referenced obj(s).
                 try:
                     related = getattr(obj, relationship, None)
-                except obj._meta.get_field_by_name(relationship)[0].model.DoesNotExist:
+                except ObjectDoesNotExist:
                     continue
                 if isinstance(related, models.Model):
                     _follow_relationships(related) 
@@ -249,8 +241,25 @@ class RevisionManager(object):
                 if self.is_registered(parent_cls):
                     parent_obj = parent_cls.objects.get(pk=obj.pk)
                     _follow_relationships(parent_obj)
-        map(_follow_relationships, object_set)
-        return result_set
+        map(_follow_relationships, object_dict)
+        # Place in the original reversions models explicitly added to the revision.
+        result_dict.update(object_dict)
+        return result_dict
+    
+    def get_version_data(self, obj, type_flag):
+        """Creates the version data to be saved to the version model."""
+        registration_info = self.get_registration_info(obj.__class__)
+        object_id = unicode(obj.pk)
+        content_type = ContentType.objects.get_for_model(obj)
+        serialized_data = serializers.serialize(registration_info.format, [obj], fields=registration_info.fields)
+        return {
+            "object_id": object_id,
+            "content_type": content_type,
+            "format": registration_info.format,
+            "serialized_data": serialized_data,
+            "object_repr": unicode(obj),
+            "type": type_flag
+        }
         
     def end(self):
         """Ends a revision."""
@@ -263,26 +272,13 @@ class RevisionManager(object):
                 if models and not self.is_invalid():
                     # Follow relationships.
                     revision_set = self.follow_relationships(models)
-                    # Because we might have uncomitted data in models, we need to 
-                    # replace the models in revision_set which might have come from the
-                    # db, with the actual models sent to reversion.
-                    diff = revision_set.difference(models)
-                    revision_set = models.union(diff)
                     # Create all the versions without saving them
                     new_versions = []
-                    for obj in revision_set:
+                    for obj, version_data in revision_set.iteritems():
                         # Proxy models should not actually be saved to the revision set.
                         if obj._meta.proxy:
                             continue
-                        registration_info = self.get_registration_info(obj.__class__)
-                        object_id = unicode(obj.pk)
-                        content_type = ContentType.objects.get_for_model(obj)
-                        serialized_data = serializers.serialize(registration_info.format, [obj], fields=registration_info.fields)
-                        new_versions.append(Version(object_id=object_id,
-                                                    content_type=content_type,
-                                                    format=registration_info.format,
-                                                    serialized_data=serialized_data,
-                                                    object_repr=unicode(obj)))
+                        new_versions.append(Version(**version_data))
                     # Check if there's some change in all the revision's objects.
                     save_revision = True
                     if self._state.ignore_duplicates:
@@ -300,8 +296,10 @@ class RevisionManager(object):
                     # Only save if we're always saving, or have changes.
                     if save_revision:
                         # Save a new revision.
-                        revision = Revision.objects.create(user=self._state.user,
-                                                           comment=self._state.comment)
+                        revision = Revision.objects.create(
+                            user = self._state.user,
+                            comment = self._state.comment,
+                        )
                         # Save version models.
                         for version in new_versions:
                             version.revision = revision
@@ -314,10 +312,18 @@ class RevisionManager(object):
         
     # Signal receivers.
         
-    def post_save_receiver(self, instance, sender, **kwargs):
+    def post_save_receiver(self, instance, created, **kwargs):
         """Adds registered models to the current revision, if any."""
         if self.is_active():
-            self.add(instance)
+            if created:
+                self.add(instance, VERSION_ADD)
+            else:
+                self.add(instance, VERSION_CHANGE)
+            
+    def pre_delete_receiver(self, instance, **kwargs):
+        """Adds registerted models to the current revision, if any."""
+        if self.is_active():
+            self.add(instance, VERSION_DELETE)
        
     # High-level revision management methods.
         
@@ -350,4 +356,3 @@ class RevisionManager(object):
         
 # A thread-safe shared revision manager.
 revision = RevisionManager()
-